chore: apply subscriber page in gdpr
